Move things around and add doc
