+++ src/GraphQL/LinkTypeQuery.php
@@ -3,7 +3,6 @@ namespace SilverStripe\Link\GraphQL;
 
 use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Type\Definition\Type as GraphqlType;
-use SilverStripe\GraphQL\Pagination\Connection;
 use SilverStripe\GraphQL\QueryCreator;
 use SilverStripe\Link\Type\Registry;
 use SilverStripe\Link\Type\Type;
