+++ templates/SilverStripe/Link/Models/Link.ss
